+++ Tests/ksqlDB.RestApi.Client.Tests/KSql/RestApi/Generators/TypeGeneratorTests.cs
@@ -15,21 +15,21 @@ public class TypeGeneratorTests
     //Arrange
 
     //Act
-    var statement = new TypeGenerator().Print(new TypeProperties<Address>());
+    var statement = new TypeGenerator().Print<Address>(new TypeProperties());
 
     //Assert
     statement.Should()
-      .Be($"CREATE TYPE {nameof(Address).ToUpper()} AS STRUCT<Number INT, Street VARCHAR, City VARCHAR>;");
+      .Be($"CREATE TYPE {nameof(Address)} AS STRUCT<Number INT, Street VARCHAR, City VARCHAR>;");
   }
 
   [Test]
   public void CreateType_WithTypeName()
   {
     //Arrange
-    var typeName = "MyType";
+    var typeName = "MYTYPE";
 
     //Act
-    var statement = new TypeGenerator().Print(new TypeProperties<Address> { EntityName = typeName });
+    var statement = new TypeGenerator().Print<Address>(new TypeProperties { EntityName = typeName });
 
     //Assert
     statement.Should().Be($"CREATE TYPE {typeName} AS STRUCT<Number INT, Street VARCHAR, City VARCHAR>;");
@@ -41,10 +41,10 @@ public class TypeGeneratorTests
     //Arrange
 
     //Act
-    var statement = new TypeGenerator().Print(new TypeProperties<Person>());
+    var statement = new TypeGenerator().Print<Person>(new TypeProperties());
 
     //Assert
-    statement.Should().Be($"CREATE TYPE {nameof(Person).ToUpper()} AS STRUCT<Name VARCHAR, Address ADDRESS>;");
+    statement.Should().Be($"CREATE TYPE {nameof(Person)} AS STRUCT<Name VARCHAR, Address ADDRESS>;");
   }
 
   [Test]
@@ -53,10 +53,10 @@ public class TypeGeneratorTests
     //Arrange
 
     //Act
-    var statement = new TypeGenerator().Print(new TypeProperties<Thumbnail>());
+    var statement = new TypeGenerator().Print<Thumbnail>(new TypeProperties());
 
     //Assert
-    statement.Should().Be($"CREATE TYPE {nameof(Thumbnail).ToUpper()} AS STRUCT<Image BYTES>;");
+    statement.Should().Be($"CREATE TYPE {nameof(Thumbnail)} AS STRUCT<Image BYTES>;");
   }
 
   [Test]
@@ -65,35 +65,35 @@ public class TypeGeneratorTests
     //Arrange
 
     //Act
-    var statement = new TypeGenerator().Print(new TypeProperties<Container>());
+    var statement = new TypeGenerator().Print<Container>(new TypeProperties());
 
     //Assert
-    statement.Should().Be($"CREATE TYPE {nameof(Container).ToUpper()} AS STRUCT<Values2 MAP<VARCHAR, INT>>;");
+    statement.Should().Be($"CREATE TYPE {nameof(Container)} AS STRUCT<Values2 MAP<VARCHAR, INT>>;");
   }
 
-  [TestCase(IdentifierEscaping.Never, ExpectedResult = "CREATE TYPE ROWTIME AS STRUCT<Value VARCHAR>;")]
-  [TestCase(IdentifierEscaping.Keywords, ExpectedResult = "CREATE TYPE ROWTIME AS STRUCT<Value VARCHAR>;")]
-  [TestCase(IdentifierEscaping.Always, ExpectedResult = "CREATE TYPE `ROWTIME` AS STRUCT<`Value` VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Never, ExpectedResult = "CREATE TYPE RowTime AS STRUCT<Value VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Keywords, ExpectedResult = "CREATE TYPE RowTime AS STRUCT<Value VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Always, ExpectedResult = "CREATE TYPE `RowTime` AS STRUCT<`Value` VARCHAR>;")]
   public string CreateType_WithSystemColumName(IdentifierEscaping escaping) =>
-    new TypeGenerator().Print(new TypeProperties<RowTime> { IdentifierEscaping = escaping });
+    new TypeGenerator().Print<RowTime>(new TypeProperties { IdentifierEscaping = escaping });
 
-  [TestCase(IdentifierEscaping.Never, ExpectedResult = "CREATE TYPE VALUES AS STRUCT<Value VARCHAR>;")]
-  [TestCase(IdentifierEscaping.Keywords, ExpectedResult = "CREATE TYPE `VALUES` AS STRUCT<Value VARCHAR>;")]
-  [TestCase(IdentifierEscaping.Always, ExpectedResult = "CREATE TYPE `VALUES` AS STRUCT<`Value` VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Never, ExpectedResult = "CREATE TYPE Values AS STRUCT<Value VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Keywords, ExpectedResult = "CREATE TYPE `Values` AS STRUCT<Value VARCHAR>;")]
+  [TestCase(IdentifierEscaping.Always, ExpectedResult = "CREATE TYPE `Values` AS STRUCT<`Value` VARCHAR>;")]
   public string CreateType_WithReservedWord(IdentifierEscaping escaping) =>
-    new TypeGenerator().Print(new TypeProperties<Values> { IdentifierEscaping = escaping });
+    new TypeGenerator().Print<Values>(new TypeProperties { IdentifierEscaping = escaping });
 
   [TestCase(IdentifierEscaping.Never,
     ExpectedResult =
-      "CREATE TYPE SYSTEMCOLUMN AS STRUCT<RowTime VARCHAR, RowOffset VARCHAR, RowPartition VARCHAR, WindowStart VARCHAR, WindowEnd VARCHAR>;")]
+      "CREATE TYPE SystemColumn AS STRUCT<RowTime VARCHAR, RowOffset VARCHAR, RowPartition VARCHAR, WindowStart VARCHAR, WindowEnd VARCHAR>;")]
   [TestCase(IdentifierEscaping.Keywords,
     ExpectedResult =
-      "CREATE TYPE SYSTEMCOLUMN AS STRUCT<RowTime VARCHAR, RowOffset VARCHAR, RowPartition VARCHAR, WindowStart VARCHAR, WindowEnd VARCHAR>;")]
+      "CREATE TYPE SystemColumn AS STRUCT<RowTime VARCHAR, RowOffset VARCHAR, RowPartition VARCHAR, WindowStart VARCHAR, WindowEnd VARCHAR>;")]
   [TestCase(IdentifierEscaping.Always,
     ExpectedResult =
-      "CREATE TYPE `SYSTEMCOLUMN` AS STRUCT<`RowTime` VARCHAR, `RowOffset` VARCHAR, `RowPartition` VARCHAR, `WindowStart` VARCHAR, `WindowEnd` VARCHAR>;")]
+      "CREATE TYPE `SystemColumn` AS STRUCT<`RowTime` VARCHAR, `RowOffset` VARCHAR, `RowPartition` VARCHAR, `WindowStart` VARCHAR, `WindowEnd` VARCHAR>;")]
   public string CreateType_WithSystemColumnNameField(IdentifierEscaping escaping) =>
-    new TypeGenerator().Print(new TypeProperties<SystemColumn> { IdentifierEscaping = escaping });
+    new TypeGenerator().Print<SystemColumn>(new TypeProperties { IdentifierEscaping = escaping });
 
   public record Address
   {
@@ -176,11 +176,11 @@ public class TypeGeneratorTests
 
     //Act
     var statement =
-      new TypeGenerator().Print(new TypeProperties<DatabaseChangeObject<IoTSensor>>());
+      new TypeGenerator().Print<DatabaseChangeObject<IoTSensor>>(new TypeProperties());
 
     //Assert
     statement.Should()
-      .Be("CREATE TYPE DATABASECHANGEOBJECT AS STRUCT<Before IOTSENSOR, After IOTSENSOR, Source SOURCE, Op VARCHAR, TsMs BIGINT>;"); //, Transaction OBJECT
+      .Be($"CREATE TYPE {nameof(DatabaseChangeObject)} AS STRUCT<Before IOTSENSOR, After IOTSENSOR, Source SOURCE, Op VARCHAR, TsMs BIGINT>;"); //, Transaction OBJECT
   }
 
   #endregion