+++ ksqlDb.RestApi.Client/KSql/RestApi/Statements/EntityInfo.cs
@@ -1,15 +1,12 @@
 using System.Reflection;
-using ksqlDB.RestApi.Client.KSql.RestApi.Enums;
-using ksqlDb.RestApi.Client.KSql.RestApi.Parsers;
 using ksqlDB.RestApi.Client.KSql.RestApi.Statements.Annotations;
-using ksqlDB.RestApi.Client.KSql.RestApi.Statements.Properties;
-using Pluralize.NET;
+using ksqlDb.RestApi.Client.KSql.RestApi.Statements.Providers;
 
 namespace ksqlDB.RestApi.Client.KSql.RestApi.Statements;
 
-internal class CreateEntityStatement
+internal class EntityInfo
 {
-  protected static readonly IPluralize EnglishPluralizationService = new Pluralizer();
+  protected static readonly EntityProvider EntityProvider = new();
 
   protected static IEnumerable<MemberInfo> Members<T>(bool? includeReadOnly = null)
   {
@@ -27,19 +24,6 @@ internal class CreateEntityStatement
       .Concat(fields);
       
     return properties.Where(c => !c.GetCustomAttributes().OfType<IgnoreByInsertsAttribute>().Any());
-  }
-
-  protected static string GetEntityName<T>(IEntityCreationProperties metadata)
-  {
-    string entityName = metadata?.EntityName;
-
-    if (string.IsNullOrEmpty(entityName))
-      entityName = typeof(T).Name;
-
-    if (metadata is { ShouldPluralizeEntityName: true })
-      entityName = EnglishPluralizationService.Pluralize(entityName);
-
-    return IdentifierUtil.Format(entityName, metadata?.IdentifierEscaping ?? IdentifierEscaping.Never);
   }
 
   protected static Type GetMemberType(MemberInfo memberInfo)