+++ ksqlDb.RestApi.Client/ChangeLog.md
@@ -1,5 +1,10 @@
 # ksqlDB.RestApi.Client
 
+# 4.0.0-rc.2
+- introduced new overloads for dropping entities: `DropTableAsync`, `DropTypeAsync`, and `DropStreamAsync` in `KSqlDbRestApiClient`. These overloads now accept an argument `DropFromItemProperties`, providing more flexibility in configuring the drop operations.
+- extracted `IKSqlDbDropRestApiClient` and `IKSqlDbCreateRestApiClient` interfaces from `IKSqlDbRestApiClient`
+- see also [breakingchanges.md](https://github.com/tomasfabian/ksqlDB.RestApi.Client-DotNet/blob/main/docs/breaking_changes.md#v400)
+
 # 3.6.2
 - `JsonPropertyName` and `PseudoColumn` attributes are taken into account when setting column names in LINQ column selection syntax #61 (contributed by @mrt181)